implement js-doc-insert-function-doc-snippet
This uses yasnippet to insert the jsdoc comment which enables the use of
placeholder values and cycling through the various fields.

To enable this, common functionality has been factored out to separate
functions, most notably js-doc--function-doc-metadata which parses the
function metadata into an alist which is then used by the jsdoc
insertion functions.