fix positioning
The insertion functions are responsible for positioning the point at the
beginning of the function by calling js-doc--beginning-of-defun, which
is necessary for js-doc--function-doc-metadata to work.

The position-compensation after the point moved around for parsing the
data that was performed with the previous implementation is no longer
necessary because the metadata parsing is done in a separate excursion
via save-excursion.